Improve B8
